+++ mods/Veil/common/src/main/java/foundry/veil/api/client/render/VeilRenderer.java
@@ -95,15 +95,6 @@ public class VeilRenderer implements ResourceManagerReloadListener {
         resourceManager.registerReloadListener(this.dynamicRenderTypeManager);
         resourceManager.registerReloadListener(this.flareEffectManager.getShellManager());
         resourceManager.registerReloadListener(this);
-
-        if (VeilRenderSystem.hasImGui()) {
-            this.registerInspectors();
-        }
-    }
-
-    @ApiStatus.Internal
-    public void registerInspectors() {
-        VeilClient.clientPlatform().onRegisterInspectors(new EditorRegistry(this.editorManager));
     }
 
     @ApiStatus.Internal
@@ -307,12 +298,5 @@ public class VeilRenderer implements ResourceManagerReloadListener {
     @Override
     public void onResourceManagerReload(@NotNull ResourceManager resourceManager) {
         VeilBloomRenderer.tryEnable();
-    }
-
-    private record EditorRegistry(EditorManager editorManager) implements VeilRegisterInspectorsEvent.Registry {
-        @Override
-        public void registerInspector(Inspector inspector) {
-            this.editorManager.add(inspector);
-        }
     }
 }